+++ libs/angular/src/lib/context-menu/context-menu.module.ts
@@ -1,0 +1,10 @@
+import { CommonModule } from '@angular/common'
+import { NgModule } from '@angular/core'
+import { NggContextMenuComponent } from './context-menu.component'
+
+@NgModule({
+  declarations: [NggContextMenuComponent],
+  imports: [CommonModule],
+  exports: [NggContextMenuComponent],
+})
+export class NggContextMenuModule {}