+++ libs/extract/src/lib/datepicker/datepicker.ts
@@ -64,6 +64,7 @@ export type DatepickerListener = (
 ) => void
 
 export interface DatepickerOptions {
+  label?: string
   locale?: string
   currentDate?: Date | string
   selectedDate?: Date | string
@@ -278,6 +279,9 @@ export const createDatepicker = (
       listener(data)
     },
     setMonth: (number) => {
+      if (typeof number !== 'number' || number < 0 || number > 11) {
+        return
+      }
       const date = new Date(data.date.setMonth(number))
       if (date < startOfMonthMinDate || date > endOfMonthMaxDate) {
         return
@@ -296,6 +300,9 @@ export const createDatepicker = (
       listener(data)
     },
     setYear: (number) => {
+      if (typeof number !== 'number') {
+        return
+      }
       const date = new Date(data.date.setFullYear(number))
       if (date < startOfMonthMinDate || date > endOfMonthMaxDate) {
         return